pyproject.toml
[project]
name = "serializd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "curl-cffi>=0.14.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
